feat: Implement Direct Construction of Minimal Acyclic Subsequential Transducers (#1)
The acyclic transducer construction described in https://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.24.3698 forms the basis of the FST used in Lucene to search for terms in the index.

This implementation does not match the implementation in Lucene. But could help found the basis for a transducer constructor that generates FSTs that can be consumed by Lucene.